Refactoring - remove anonymous namespace in header.
If you implement a function in header and you want to avoid from
multiple declaration, you should use `inline` keyword on the function
declaration.

`anonymous namespace`, technically speaking, just make `static` prefix
on function declarations. `static` keyword ensures compilers to make
sections for every functions, but the visibility of those functions will
are `weak`. `weak` functions just make the same effect on ELF, but it's
a wrong approach to put implementation on header if you are not using an
ELF system.

In case of using ELF, `inline` keyword makes multiple COMDAT sections and
linkers is asked to resolve them as one. `inline` and COMDAT sections
are the correct approach to put implementation on headers for any
system.